Added the environment variables for tanent and client Id.
